NF: move Undo task to UndoAction
